finsih point low memory query

+++ tests/test_framework/test_config.h
@@ -8,7 +8,7 @@
 using Coord = long;
 // using Coord = double;
 #else
-using Coord = long;
+using Coord = unsigned long long;
 // using Coord = double;
 #endif  // CCP
